requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
psycopg2-binary==2.9.11
pyodbc==5.2.0
#face-recognition==1.3.0
opencv-python==4.8.1.78
#numpy==1.24.3
numpy>=2.2,<2.5
#Pillow==10.1.0
Pillow>=11.0.0
cryptography==41.0.7
python-dotenv==1.0.0
werkzeug==3.0.1
bcrypt==4.1.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
mediapipe>=0.10.14
onnxruntime>=1.20.0
#python-3.11.8
pyodbc==5.2.0
